backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
aiohttp==3.9.1
